Swap final node after exiting loop
By moving `next` to the loop condition we exit before accessing a null
value `next.next` (would not be able to find property of null when on
tail, since D->null), however, D still needs to point to the last prev
value C (the final swap); so we do it after exiting the loop, assign the final node to
the lists head, then return the list.

Note the previous version of this file runs fine, this commit is for
concision.